feat:상세게시글과 수정하기 페이지 연결

+++ src/components/units/board/detail/BoardDetail.container.js
@@ -10,23 +10,38 @@ export default function BoardDetail() {
     variables : {boardId : router.query.boardId},
   });
   const [ deleteBoard ] = useMutation(DELETE_BOARD);
-
+  
   const onClickDeleteBoard = async (e) => {
-    confirm("삭제하시겠습니까?")
-    try {
-    if(e.target.id){
-    const res = await deleteBoard({
-      variables : {ID : e.target.id} 
-      })
-      }} catch(err) {
-      console.log(err)
+    const answer = confirm("삭제하시겠습니까?")
+    if (answer) {
+      try {
+      if(e.target.id){
+      const res = await deleteBoard({
+        variables : {ID : e.target.id} 
+        })
+        }} catch (err) {
+        console.log(err)
+      }
+      router.push('/boards')
+    } else {
+      e.preventDefault();
     }
-    router.push('/boards')
   }
+
+  const onclickMoveToEdit = () => {
+    console.log("클릭")
+    router.push(`/boards/${router.query.boardId}/edit`);
+  }
+
+  const onClickMoveToList = () => {
+    router.push('/boards');
+  } 
 
   return (
     <BoardDetailUI data = {data}
     onClickDeleteBoard = {onClickDeleteBoard}
+    onclickMoveToEdit = {onclickMoveToEdit}
+    onClickMoveToList = {onClickMoveToList}
     />
   )
 }
